+++ lib/rdstation/contacts.rb
@@ -3,9 +3,10 @@ module RDStation
   # More info: https://developers.rdstation.com/pt-BR/reference/contacts
   class Contacts
     include HTTParty
-    
-    def initialize(authorization_header:)
-      @authorization_header = authorization_header
+    include ::RDStation::RetryableRequest
+
+    def initialize(authorization:)
+      @authorization = authorization
     end
 
     #
@@ -13,13 +14,17 @@ module RDStation
     #   The unique uuid associated to each RD Station Contact.
     #
     def by_uuid(uuid)
-      response = self.class.get(base_url(uuid), headers: @authorization_header.to_h)
-      ApiResponse.build(response)
+      retryable_request(@authorization) do |authorization|
+        response = self.class.get(base_url(uuid), headers: authorization.headers)
+        ApiResponse.build(response)
+      end
     end
 
     def by_email(email)
-      response = self.class.get(base_url("email:#{email}"), headers: @authorization_header.to_h)
-      ApiResponse.build(response)
+      retryable_request(@authorization) do |authorization|
+        response = self.class.get(base_url("email:#{email}"), headers: authorization.headers)
+        ApiResponse.build(response)
+      end
     end
 
     # The Contact hash may contain the following parameters:
@@ -34,8 +39,10 @@ module RDStation
     # :website
     # :tags
     def update(uuid, contact_hash)
-      response = self.class.patch(base_url(uuid), :body => contact_hash.to_json, :headers => @authorization_header.to_h)
-      ApiResponse.build(response)
+      retryable_request(@authorization) do |authorization|
+        response = self.class.patch(base_url(uuid), :body => contact_hash.to_json, :headers => authorization.headers)
+        ApiResponse.build(response)
+      end
     end
 
     #
@@ -47,14 +54,16 @@ module RDStation
     #   Contact data
     #
     def upsert(identifier, identifier_value, contact_hash)
-      path = "#{identifier}:#{identifier_value}"
-      response = self.class.patch(base_url(path), body: contact_hash.to_json, headers: @authorization_header.to_h)
-      ApiResponse.build(response)
+      retryable_request(@authorization) do |authorization|
+        path = "#{identifier}:#{identifier_value}"
+        response = self.class.patch(base_url(path), body: contact_hash.to_json, headers: authorization.headers)
+        ApiResponse.build(response)
+      end
     end
 
     private
 
-    def base_url(path = "")
+    def base_url(path = '')
       "https://api.rd.services/platform/contacts/#{path}"
     end
   end