+++ lib/rdstation/fields.rb
@@ -3,16 +3,19 @@ module RDStation
   # More info: https://developers.rdstation.com/pt-BR/reference/contacts
   class Fields
     include HTTParty
+    include ::RDStation::RetryableRequest
 
     BASE_URL = 'https://api.rd.services/platform/contacts/fields'.freeze
     
-    def initialize(authorization_header:)
-      @authorization_header = authorization_header
+    def initialize(authorization:)
+      @authorization = authorization
     end
 
     def all
-      response = self.class.get(BASE_URL, headers: @authorization_header.to_h)
-      ApiResponse.build(response)
+      retryable_request(@authorization) do |authorization|
+        response = self.class.get(BASE_URL, headers: authorization.headers)
+        ApiResponse.build(response)
+      end
     end
 
   end