+++ spec/lib/rdstation/events_spec.rb
@@ -6,13 +6,13 @@ RSpec.describe RDStation::Events do
   let(:expired_access_token) { 'expired_access_token' }
 
   let(:event_with_valid_token) do
-    described_class.new(authorization_header: RDStation::AuthorizationHeader.new(access_token: valid_access_token))
+    described_class.new(authorization: RDStation::Authorization.new(access_token: valid_access_token))
   end
   let(:event_with_expired_token) do
-    described_class.new(authorization_header: RDStation::AuthorizationHeader.new(access_token: expired_access_token))
+    described_class.new(authorization: RDStation::Authorization.new(access_token: expired_access_token))
   end
   let(:event_with_invalid_token) do
-    described_class.new(authorization_header: RDStation::AuthorizationHeader.new(access_token: invalid_access_token))
+    described_class.new(authorization: RDStation::Authorization.new(access_token: invalid_access_token))
   end
 
   let(:events_endpoint) { 'https://api.rd.services/platform/events' }
@@ -106,6 +106,11 @@ RSpec.describe RDStation::Events do
           'reason' => 'Lost reason'
         }
       }
+    end
+
+    it 'calls retryable_request' do
+      expect(event_with_valid_token).to receive(:retryable_request)
+      event_with_valid_token.create({})
     end
 
     context 'with a valid auth token' do